Rewritten scenario 1 continuous test
Implemented telegraf reporting

Removed unused files
