--- library
@@ -1,1 +1,0 @@
-.ansible-modules/otc
