--- playbooks/library
@@ -1,1 +1,0 @@
-../library
