--- pyproject.toml
@@ -1,19 +1,0 @@
-[tool.poetry]
-name = "csm-sandbox"
-version = "0.1.0"
-description = "Sandbox for OTC customer service monitoring"
-authors = ["OTC Customer service monitoring"]
-license = "Apache-2.0"
-
-[tool.poetry.dependencies]
-python = "^3.7"
-
-jinja2 = "^2.10"
-ocomone = "^0.3.1"
-requests = "^2.22"
-
-[tool.poetry.dev-dependencies]
-
-[build-system]
-requires = ["poetry>=0.12"]
-build-backend = "poetry.masonry.api"
