+++ requirements.txt
@@ -1,0 +1,5 @@
+ocomone >= 0.3.1
+jinja2 >= 2.10
+requests >= 2.22
+pytelegraf >= 0.3.3
+wrapt == 1.11.2
